tasks/service: implement first methods with a Postgres database.
This is somewhat complete, or at least complete enough to get us started. The
tests require setting an environment variable `$TASKS_TEST_DATABASE` to a
connection string used to open up the connection to the Postgres database. The
database is assumed to have the schema in tasks/postgres/schema.sql.