docker/dockertest: package to run Docker containers in tests.
This package is intended to be used in integration tests. It provides some
building blocks for loading images, running containers, and connecting to
containers. For better ergonomics when running specific types of containers,
such as databases, it's probably a better idea to create a new package that
wraps the functions in this package.