tasks/service: add a testClient to the integration test.
The purpose of the test client is to make it easier to create sequences of calls
to the API without having to do lots of manual error checking (instead, the
error checking is done in the *T methods). This makes it easier to write tests
for more complex scenarios without having to sacrifice correctness.

One could probably set up code generation to do this instead, but for the
current API with few methods, writing it down manually is fine.