tasks: define, test, and implement ListTasks.
Implementing ListTasks means implementing pagination. With pagination, there are
quite a few things that can go wrong. There is lots to read about pagination on
https://google.aip.dev/132 and https://google.aip.dev/158 for more details.

The solution chosen for this simple service is to simply let each task have a
unique, increasing ID. Page tokens are UUIDs that map to a minimum ID to fetch
in the next page. This makes the tokens opaque and protects from the user
tampering with them. At least I think so.

Using a unique, increasing ID means that I changed the resource ID for tasks
from UUIDs to integer IDs. I think this is fine -- revealing the number of tasks
is not very sensitive for this service. After all, this service is only intended
for me to use for some personal stuff, maybe. If that is the case, one might
wonder who the "user" that could tamper with page tokens would be. The answer
is: no-one. But it's a fun exercise in service design to implement it.