postgres: package for connecting to Postgres databases.
The purpose of this package is to provide some sane defaults for retrying
connections, adding logging, and possibly other features in the future. By
embedding the `*pgxpool.Pool` type into the `Pool` type, and forcing callers to
use `Pool`, we get some help from the compiler to make sure these sane defaults
are actually used, while keeping compatibility with the `pgx` interface.

This change also includes a simple log adapter for `glog`, which is always
installed in any call to `Open`.