tasks: define, test, and implement UpdateTask.
Woah, this was a doozy. There is probably some way one can abstract over this --
there is a lot of conceptual repetition in this code, constantly checking that
we only update `title` or `description` for example -- but for now it works
well. All the tests are passing and the code is reasonably straightforward.