bazel: pass absolute path to `make` and `exec`.
If the `_tools` directory doesn't exist, this script doesn't work because `make`
only knows how to build `/absolute/path/to/_tools/bazelisk`, not
`_tools/bazelisk` relative to the root directory. I always (well, not 100 % of
the time, but 99.99 %) run the `bazel` script from the root directory, so using
`$(pwd)` isn't a very big problem.