tasks/fake: add an in-memory fake of the Tasks service.
This change does not really fulfil any purpose except for me playing around.
Having a lightweight but complete in-memory fake could potentially prove useful
in the future for integration tests or similar.

The fake is tested by running it through `tasks/testsuite`. That means it passes
the same tests as the real implementation does.