tools.mk: make the _tools directory an order-only dependency.
I don't fully understand the concept of order-only dependencies, but it
generally seems to work for the use case "make sure this directory exists before
putting stuff in it". If I had to guess, I'd say that specifying it as a regular
dependency means that any time the directory is changed -- e.g., a file within
it is created/deleted/changed -- then everything depending on the directory
needs to be updated. That's not really what I want.