tasks/fake/server: command to run the fake as a gRPC server.
I consider this to be a starting point for actually running a container
somewhere I can reach it with a client over the network. Running without any
auth etc is probably not a good idea, and someone can probably DoS the server by
spamming lots of `ListTasks` RPCs, but eh, I don't care.