tasks: add hierarchy to tasks.
This includes changing many of the RPCs to account for the possibility of
ancestor or descendant tasks. The test suite has been updated to reflect that,
and both the in-memory and Postgres implementation have been updated to pass the
test suite.

For the fake implementation, an abstraction was added to map task names to their
locations in the fake.Fake.tasks array. This makes some of these operations
easier to make because we now have to fetch tasks by name frequently.

For the Postgres implementation, the schema itself is unchanged but it includes
a bunch of views that use recursive CTEs (common table expressions) to provide a
SELECT-friendlier way to query for transitive ancestors and descendants.