tasks: add `update_time` to tasks.
This also updates the test suite to check for it. The `update_time` field is
only set when an actual change happens -- no-op updates are not treated as
updates.

Writing this feature exposed a few bugs in the Postgres implementation, which
have now been fixed (and the tests should cover them too).