Install a Go toolchain using Make.
A Go toolchain is used to install other tools such as `bazelisk` and `protoc`
plugins. Previously, a Go toolchain was required to be installed manually, which
has some drawbacks:

*   The version can differ between machines.
*   It adds an extra implicit dependency to use this repository.

This change pulls down a Go toolchain from the Go website and uses that to build
the aforementioned tools, which results in a more reproducible build.