auth/n/basic: add HTTP Basic authentication.
HTTP Basic authentication means sending username and password on every request.
This is very simple and works well for simple purposes when combined with HTTPS.

The authentication is implemented as a basic.Credentials type, which satisfies
the credentials.PerRPCCredentials interface, and a basic.Interceptor that can be
used server-side to verify incoming basic.Credentials.