grpctest: serve with TLS security by generating a certificate at build time.
This is accomplished by a new tool called `generate_certificate`. The tool is
invoked in a Bazel `genrule` to provide a certificate for the `grpctest` package
at build time. Therefore we don't have to commit the certificates to source
control, or remember to regenerate them if we ever want to change something.

One could wonder if the `grpctest` package shouldn't instead provide a way to
specify whether TLS should be used or not. I think that if you don't care then
using TLS isn't going to bother you, and if you do care then it's probably going
to be because you want TLS, not because you _don't_ want TLS.