docker/config: client configuration for Docker rules.
This will eventually be used to push container images to GCP. The `gcloud`
command line tool can be used to authenticate the Docker client, and the
`config.json` file is the way to do it. But since that contains details about
the GCP environment, it needs to be generated and then _not_ stored in version
control.

It's a bit of a convoluted setup but it's a good way to prevent myself from
shooting myself in the foot by committing things to this repository that I later
might decide should be considered "secret" information.