docker/dockertest: use a simple HTTP server instead of Postgres in tests.
Previously the test would pass when run in isolation, but often fail when
running with e.g. `--runs_per_test=100`. Which is weird, because the test should
never have passed -- we started the Postgres container the wrong way by not
supplying the environment variable `POSTGRES_PASSWORD`.

To fix both of these issues the test now uses a simple HTTP server returning
"Hello, World!".