tasks/testsuite: define a set of tests for project operations.
These are very similar to the tests for task operations, but then the resources
are very similar in shape so it makes sense.

For now, all the tests in projects.go are skipped. This allows us to have tests
that compile but are not executed. Once backends (the fake one and the Postgres
one) start implementing methods, the tests can be enabled in an incremental
fashion.