adventofcode/cc/tools: add code generator for benchmark programs.
This is heavily pattern-matched off of the generator for unit tests.

This change also includes a BUILD macro for generating benchmarks. I figured it
might as well live in the same .bzl file as the `cc_aoc_test` macro as they are
so similar, but then I had to rename the .bzl file.