adventofcode/cc: add solution for 2022/24.
This was a fun problem, and I'm happy I came up with a "closed form" solution
for calculating the positions of the blizzards, rather than having to actually
move them around. Made the rest of the code a bit easier to implement as well!

This is probably not a very fast solution, but it's sufficiently fast.

Benchmarks from `./bazel run -c opt //adventofcode/cc/year2022:day24_benchmark`:

    2023-06-17T18:33:21+01:00
    Running /home/saser/.cache/bazel/_bazel_saser/06ad534583e887506ca6aa175f8ed7b5/execroot/code/bazel-out/k8-opt/bin/adventofcode/cc/year2022/day24_benchmark
    Run on (16 X 4679.3 MHz CPU s)
    CPU Caches:
    L1 Data 32 KiB (x16)
    L1 Instruction 32 KiB (x8)
    L2 Unified 512 KiB (x8)
    L3 Unified 16384 KiB (x1)
    Load Average: 1.06, 1.20, 0.76
    ***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
    -----------------------------------------------------------------
    Benchmark                       Time             CPU   Iterations
    -----------------------------------------------------------------
    BM_Part1/day24.real.in   21318845 ns     21316569 ns           33
    BM_Part2/day24.real.in   47708781 ns     47705468 ns           14