Bump Bazel version to 6.0.0.
Staying up to date with versions. The most exciting feature in Bazel 6.0.0 is
Bzlmod, but I'm probably not going to use that for a while -- I'll let the dust
settle a little bit, let others work out various wrinkles, and wait for the
Bazel Central Registry to mature a bit before I try using it.