Upgrade several Go-related tools and dependencies.
The main part of this change is to update `rules_go`, `bazel-gazelle`, and the
Go version.

However, it seems that with Go 1.20 _something_ has changed in relation to
`glibc`. This caused problems for the `dockertest` test: it relied on a very
basic Go binary containing a HTTP server. With the `glibc` change, that HTTP
server could no longer start up, because it somehow depended on `glibc`.

To solve that, I needed some other way to start up a HTTP server. I figured an
Nginx container would be somewhat similarly lightweight.

When I pulled that in, the test `TestAddress` in `dockertest_test.go` started
failing. The reason is that Nginx doesn't start up quickly enough within the
container to start listening on port 80, and so the `http.Get` calls failed with
a "connection reset by peer" error. To work around that, I added some retries.