Migrate from `WORKSPACE` to `MODULE.bazel`.
This represents a pretty significant amount of work to move away from the old
and deprecated `WORKSPACE` way of managing Bazel external dependencies, to using
`MODULE.bazel` aka Bzlmod.

At this point, I have `bazel build //...` and `bazel test //...` up and running
smoothly. The only thing I am aware of that doesn't work is `bazel run
@hedron_compile_commands//:refresh_all`, but it seems that the issue I am
running into is getting fixed
[soon](https://github.com/hedronvision/bazel-compile-commands-extractor/pull/262).

In the process of doing this I did a few things to get the migration over the
line.

**Migrated from `rules_docker` to `rules_oci`.** The former is more or less
deprecated and there's no Bazel module for it in BCR. There is a module for
`rules_oci` though, which seems to be the spiritual successor to `rules_docker`.
They do mostly the same things, although in different ways.

**Bumped version of most dependencies.** The main reason for doing so was to get
on versions of dependencies that likely had much better support for Bzlmod.

**Struggled a lot to get the dependencies on the Docker Go libraries working**,
and ended up refactoring my old version of `dockertest` in favor of one based on
`github.com/ory/dockertest/v3` which I know from professional experience works
perfectly fine and doesn't require me to do some of the lower-level Docker stuff
myself.

**Removed a bunch of the existing Docker-related stuff**, since I couldn't be
bothered to fix it and will likely not be using it in the near future anyway.

**Removed the `./bazel` script**, and chose to just rely on Bazelisk being
installed on the system instead.

Probably more things I'm forgetting, but they're probably not that important if
I can't think of them right now.